+++ packages/design-system/src/components/InputField/Input.js
@@ -7,7 +7,7 @@ import getMergedClassNames from '../../utils/getMergedClassNames';
 const cx = classNames.bind(styles);
 
 const Input = React.forwardRef((props, ref) => {
-  const { error, className, ...restProps } = props;
+  const { error, className, width, style, ...restProps } = props;
 
   const mergedClassNames = getMergedClassNames(
     cx({
@@ -17,13 +17,20 @@ const Input = React.forwardRef((props, ref) => {
     className
   );
 
-  return <input ref={ref} className={mergedClassNames} {...restProps} />;
+  const mergedStyle = (style || width) ? {
+    width,
+    ...(style || {})
+  } : void 0;
+
+  return <input ref={ref} className={mergedClassNames} style={mergedStyle} {...restProps} />;
 });
 
 Input.propTypes = {
   error: PropTypes.string,
   className: PropTypes.string,
-  type: PropTypes.string
+  type: PropTypes.string,
+  maxLength: PropTypes.number,
+  width: PropTypes.string
 };
 
 Input.defaultProps = {