+++ packages/design-system/src/components/TextAreaField/TextAreaField.js
@@ -11,8 +11,10 @@ const TextAreaField = React.forwardRef((props, ref) => {
     labelText,
     labelAdornment,
     className,
+    textareaWidth,
     id,
     fieldClassName,
+    labelRightNode,
     ...restProps
   } = props;
 
@@ -25,10 +27,12 @@ const TextAreaField = React.forwardRef((props, ref) => {
       labelAdornment={labelAdornment}
       className={className}
       htmlFor={id}
+      labelRightNode={labelRightNode}
     >
       <TextArea
         id={id}
         ref={ref}
+        width={textareaWidth}
         {...restProps}
         className={fieldClassName}
         error={error}
@@ -44,9 +48,12 @@ TextAreaField.propTypes = {
   id: PropTypes.string.isRequired,
   className: PropTypes.string,
   inline: PropTypes.bool,
+  /** use to easily setup textarea width - by default the element takes 100% of its container width */
+  textareaWidth: PropTypes.string,
   error: PropTypes.string,
   description: PropTypes.node,
-  fieldClassName: PropTypes.string
+  fieldClassName: PropTypes.string,
+  labelRightNode: PropTypes.node
 };
 
 export default TextAreaField;