+++ packages/design-system/src/components/TextField/TextField.js
@@ -1,11 +1,11 @@
-import * as React from 'react';
-import * as PropTypes from 'prop-types';
-import classNames from 'classnames/bind';
-import styles from './style.scss';
-import FieldLabel from '../FieldLabel';
-import FieldError from '../FieldError';
-import FieldDescription from '../FieldDescription';
-import getMergedClassNames from '../../utils/getMergedClassNames';
+import * as React from "react";
+import * as PropTypes from "prop-types";
+import classNames from "classnames/bind";
+import styles from "./style.scss";
+import FieldLabel from "../FieldLabel";
+import FieldError from "../FieldError";
+import FieldDescription from "../FieldDescription";
+import getMergedClassNames from "../../utils/getMergedClassNames";
 
 const cx = classNames.bind(styles);
 
@@ -18,10 +18,11 @@ const TextField = props => {
     labelAdornment,
     className,
     htmlFor,
-    children
+    children,
+    labelRightNode    
   } = props;
 
-  const baseClass = 'text-field';
+  const baseClass = "text-field";
   const mergedClassNames = getMergedClassNames(
     cx({
       [baseClass]: true,
@@ -32,25 +33,41 @@ const TextField = props => {
 
   return (
     <div className={mergedClassNames}>
-      {labelText && (
-        <div
-          className={cx({
-            [`${baseClass}__label`]: true,
-            [`${baseClass}__label--inline`]: inline
-          })}
-        >
-          <FieldLabel htmlFor={htmlFor}>{labelText}</FieldLabel>
-          {labelAdornment && (
-            <div className={cx({ [`${baseClass}__label-adornment`]: true })}>
-              {labelAdornment}
-            </div>
-          )}
-        </div>
+      {labelRightNode && inline && (
+        <React.Fragment>
+          <div className={cx([`${baseClass}__label-right-node`, `${baseClass}__label-right-node--inline`])}>
+            {labelRightNode}
+          </div>
+          <div className={cx(`${baseClass}__row-break`)}></div>
+        </React.Fragment>
       )}
-      <div>
-        {children}
-        {error && <FieldError>{error}</FieldError>}
-        {description && <FieldDescription>{description}</FieldDescription>}
+      <div className={cx([`${baseClass}__wrapper`, inline && `${baseClass}__wrapper--inline`])}>
+        {(labelText || labelRightNode ) && (
+          <div className={cx([`${baseClass}__label`, inline && `${baseClass}__label--inline`, !labelText && `${baseClass}__label--no-text`])}>
+            {labelText && (
+              <div
+                className={cx(`${baseClass}__label-wrapper`)}
+              >
+                <FieldLabel htmlFor={htmlFor}>{labelText}</FieldLabel>
+                {labelAdornment && (
+                  <div className={cx(`${baseClass}__label-adornment`)}>
+                    {labelAdornment}
+                  </div>
+                )}
+              </div>
+            )}
+            {labelRightNode && !inline && (
+              <div className={cx(`${baseClass}__label-right-node`)}>
+                {labelRightNode}
+              </div>
+            )}
+          </div>
+        )}
+        <div className={cx(`${baseClass}__content`)}>
+          {children}
+          {error && <FieldError>{error}</FieldError>}
+          {description && <FieldDescription>{description}</FieldDescription>}
+        </div>
       </div>
     </div>
   );
@@ -65,7 +82,8 @@ TextField.propTypes = {
   inline: PropTypes.bool,
   error: PropTypes.string,
   description: PropTypes.node,
-  children: PropTypes.node
+  children: PropTypes.node,
+  labelRightNode: PropTypes.node
 };
 
 export default TextField;