+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,15 @@
 All notable changes to this project will be documented in this file.
 
+## v0.13.0
+
+_Release: 2021-06-11_
+
+##### New features
+
+- adding new labelRightNode property to InputField and TextAreaField components #234 #250
+- adding new properties inputWidth and textareaWidth properties to change the default width value #250
+- changing default value of Input and TextArea width to 100% - it's a breaking change #250
+
 ## v0.12.3
 
 _Release: 2021-06.01_